+++ docs/api/skills/abstract_round_abci/tests/test_tools/base.md
@@ -1,0 +1,78 @@
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base"></a>
+
+# packages.valory.skills.abstract`_`round`_`abci.tests.test`_`tools.base
+
+Tests for abstract_round_abci/test_tools/common.py
+
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base.FSMBehaviourTestToolSetup"></a>
+
+## FSMBehaviourTestToolSetup Objects
+
+```python
+class FSMBehaviourTestToolSetup()
+```
+
+BaseRandomnessBehaviourTestSetup
+
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base.FSMBehaviourTestToolSetup.setup_class"></a>
+
+#### setup`_`class
+
+```python
+@classmethod
+def setup_class(cls) -> None
+```
+
+Setup class
+
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base.FSMBehaviourTestToolSetup.teardown_class"></a>
+
+#### teardown`_`class
+
+```python
+@classmethod
+def teardown_class(cls) -> None
+```
+
+Teardown class
+
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base.FSMBehaviourTestToolSetup.setup"></a>
+
+#### setup
+
+```python
+def setup() -> None
+```
+
+Setup test
+
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base.FSMBehaviourTestToolSetup.teardown"></a>
+
+#### teardown
+
+```python
+def teardown() -> None
+```
+
+Teardown test
+
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base.FSMBehaviourTestToolSetup.set_path_to_skill"></a>
+
+#### set`_`path`_`to`_`skill
+
+```python
+def set_path_to_skill(path_to_skill: Path = PATH_TO_SKILL) -> None
+```
+
+Set path_to_skill
+
+<a id="packages.valory.skills.abstract_round_abci.tests.test_tools.base.FSMBehaviourTestToolSetup.setup_test_cls"></a>
+
+#### setup`_`test`_`cls
+
+```python
+def setup_test_cls(**kwargs: Any) -> FSMBehaviourBaseCase
+```
+
+Helper method to setup test to be tested
+
