+++ packages/packages.json
@@ -7,31 +7,31 @@
         "protocol/valory/tendermint/0.1.0": "bafybeid6b2wlb24g6d3godmqms44qvnpkhlvb27icotuobvnscmdmlhaha",
         "skill/valory/abstract_abci/0.1.0": "bafybeicdiwkvotgr4h2zulx5rsgxqid22d5greqqhgqadd2uqlmpioipem",
         "contract/valory/gnosis_safe/0.1.0": "bafybeiem74jz5qgzkxwt34wbdlo64cloaerjui2jme4svdtjx4nh6v2d2y",
-        "skill/valory/abstract_round_abci/0.1.0": "bafybeiah3wy5fffoidqpxhvfhnmjqx2ohquex7jhpzyvca522ncsbtaqki",
+        "skill/valory/abstract_round_abci/0.1.0": "bafybeihlojtrgm5vufp5iioz6yiykdaqhlrxxgkzn6w624ze2gqzhmalfq",
         "contract/valory/multisend/0.1.0": "bafybeigjywkl7hydjsrkogob3xebj2ifhqwmfhhxoeyrndzhhxi5u6amey",
-        "skill/valory/transaction_settlement_abci/0.1.0": "bafybeih5blwasq5ulbp4ymifii6c2ozeujewraosg7ytgsosbbeesxe3xi",
-        "skill/valory/registration_abci/0.1.0": "bafybeifd234ranbtdh7yepyzvlig5yertcnhsjedsmxz3ec5mjr4oajcde",
-        "skill/valory/reset_pause_abci/0.1.0": "bafybeihffxykgdylhk7qlo2lrvktw2tygyx6n2hlvcaytkgncviofclob4",
-        "skill/valory/termination_abci/0.1.0": "bafybeia2coq7bc6adztmf46a2v63pl3uayyu3lmss7onxplqfgbqjj74wa",
+        "skill/valory/transaction_settlement_abci/0.1.0": "bafybeifbn4b3g6ogeo2xwrtnou27hkpag3meuf5fxkaax523xwekksnqtu",
+        "skill/valory/registration_abci/0.1.0": "bafybeibpktkfwnkxcz2natvrwssoo5diqgchiolv5a42f3gscwwzkgo5z4",
+        "skill/valory/reset_pause_abci/0.1.0": "bafybeiely5anbuehpojf27jbalecam5qwkg7ltf2asjkg4ep4l7zdedhye",
+        "skill/valory/termination_abci/0.1.0": "bafybeidar6ao7jcplieycdqyhsaee3nvqnuu52wiztnsmxgkto5gbpjgna",
         "skill/valory/counter/0.1.0": "bafybeibwg3lkuj536tm7njlbfuslamsoqes62oshp3ad3fhlomqfk3wugu",
         "skill/valory/counter_client/0.1.0": "bafybeic7qlw7vyovllmu35rb3cag4afduemo6ulr7sfkxtwtrjhlb2a5cq",
-        "skill/valory/hello_world_abci/0.1.0": "bafybeib5pgg6pgwd6l4ogjdvyfimbdqx4p3zdl2vxtjj2lyy2rql2lfmze",
-        "skill/valory/register_reset_abci/0.1.0": "bafybeiczrjsuscfn5dsvmsifwik4uvesssz4qp5m6rd72lqfxwbqe2fpy4",
-        "skill/valory/register_termination_abci/0.1.0": "bafybeihs6lfflm3dbdcworyuns3uadpivrl7xwqszjx7doc7c53kfojucu",
-        "skill/valory/test_abci/0.1.0": "bafybeidgb6hxxtsgegndnc4sgu4zfexfopwdr5ipokf4mewglm6jihklre",
+        "skill/valory/hello_world_abci/0.1.0": "bafybeibte3tkdzrftrvuh42ierntgp3nqislfoznucdddsgjq3lqagxoou",
+        "skill/valory/register_reset_abci/0.1.0": "bafybeidrsesudatsuqgsj4xqcf64yjciqklnlzdbie6vdzq5e2ft56npja",
+        "skill/valory/register_termination_abci/0.1.0": "bafybeic7hoiwsq4k5yz767zgqqp7zehwe435qdquo4sypnmv6b5vr73fta",
+        "skill/valory/test_abci/0.1.0": "bafybeie3fxmx6snuwmqrlaorinnxnnwra5vbevl4c4vf4wjjfr3auydlty",
         "agent/valory/abstract_abci/0.1.0": "bafybeibet7hs6ihznaiqy4wfwbbhdadqz7uc6m6mbnofttcswr54szg7rq",
         "agent/valory/counter/0.1.0": "bafybeifap4p65f5pyee2ij7ng4xxje6vswutrmy7cszpfu4kmk5ehx2qwe",
         "agent/valory/counter_client/0.1.0": "bafybeic4crnn7t5ep6bsjrdste2tdbwoktftdhtooxeweypmujv5t7cmpi",
-        "agent/valory/hello_world/0.1.0": "bafybeig3uz5zjb25xbx7uvlj2d5d4d74vilb6byj5mwrdkk4dxtdx7ezya",
-        "agent/valory/register_reset/0.1.0": "bafybeidyjck5jf5dg4mwid3ysfpupw24vbgook54efn3xvhcb5dy5t7z5y",
-        "agent/valory/register_termination/0.1.0": "bafybeidj2mvuhpuxs5bqtmidxc74eut3bd6o5hsiwl5of6u4vocrcsgzre",
-        "agent/valory/registration_start_up/0.1.0": "bafybeiajx3u4llj6pxo6fzgt4oyqkg22hfpqtva45isghgjvwnkmy3flpy",
-        "agent/valory/test_abci/0.1.0": "bafybeigbakxexwwe6o5afn3qaytfrf2n6i5md62ibu4a2t5emeh2vuw6cy",
+        "agent/valory/hello_world/0.1.0": "bafybeibijwr4jbcaxuthao3iuph77jfenomtooqmgtg5lyrpo72pvwil3m",
+        "agent/valory/register_reset/0.1.0": "bafybeigxvsvwstrjijefmrigfox3vgy7hpdway4tdtn2qzgubjjgj6anuq",
+        "agent/valory/register_termination/0.1.0": "bafybeicacmvljf6r24odef3vggfp7n4zer2sdwl3x5wambrbd2aphz5kba",
+        "agent/valory/registration_start_up/0.1.0": "bafybeihvniftqplc4clesyo5xcmm4baws3bmxggw2uwgjykorbnv4jvdpi",
+        "agent/valory/test_abci/0.1.0": "bafybeifuzvyp6k2jtanpt3z4yhlychi7f4nuh4c5k6ydgqyol4ap23aybi",
         "service/valory/counter/0.1.0": "bafybeiasglbtxdxt7tuehgmyfuu3726yncfnlfcznensz7iz6mqozqpnqm",
-        "service/valory/hello_world/0.1.0": "bafybeifbol2cioamgjmkn6dibfqgaeg5oyzyx3755dk7f6mpg4gqikjvyu",
-        "service/valory/register_reset/0.1.0": "bafybeicbzqmjor3tuzhmt32ppw5j23nb5unzec2d5wiuldd53gj2e2tug4",
-        "skill/valory/register_reset_recovery_abci/0.1.0": "bafybeie74fu6oxg7icw4vojnmvodb6qb3pbyqrzmmxmt5hnevfindcykfa",
-        "agent/valory/register_reset_recovery/0.1.0": "bafybeicheafg64vonxnvaswv3ictwxv36zk5agzcpdwd4pyweefttuyvde"
+        "service/valory/hello_world/0.1.0": "bafybeicfxd5ed272yy265c5olbtosjpzvatl6l5hisrqzxr52ygwjxmfba",
+        "service/valory/register_reset/0.1.0": "bafybeidbopjztk7nzkol7qvxcfp66ydfsyayjepdtat3tlumtgol67ibnm",
+        "skill/valory/register_reset_recovery_abci/0.1.0": "bafybeifnz7vdpj5zdh36xuhvatwrmtvx2shuchb3a5oybqfga2zf37f2rm",
+        "agent/valory/register_reset_recovery/0.1.0": "bafybeibyke4tmo4l2ktrhchh24ertve5emofruk7bu4rfllcthl5qzv3za"
     },
     "third_party": {
         "protocol/open_aea/signing/1.0.0": "bafybeiambqptflge33eemdhis2whik67hjplfnqwieoa6wblzlaf7vuo44",
