+++ packages/valory/skills/abstract_round_abci/skill.yaml
@@ -53,10 +53,10 @@ fingerprint:
   tests/test_models.py: bafybeia6yf46ver2qoxn2km4pqlbrv2ng76c2stgcyrdzx66e2c6r32sne
   tests/test_serializer.py: bafybeiddbdxrr2nhzemk6jrcxmhjwwfxur6dfj3kpeyrgme7xdggf54p6m
   tests/test_tools/__init__.py: bafybeiew6gu4pgp2sjevq4dbnmv2ail5dph7vj4yi7h3eae4gzx7vj7cbq
-  tests/test_tools/helpers.py: bafybeia3sbvrvidzwgmqbqzqi5lbmhs7pnzcmcxbriduc7np5o552ckcle
-  tests/test_tools/test_base.py: bafybeiagjvpgrzrackzpk6nvxj32vzg2skrpavwvwqjpflnqkq226nrfpa
-  tests/test_tools/test_common.py: bafybeifytdvuvf4q2yg74ctnj2jygdjmpcaejo23t6e65rsyimk6njzxca
-  tests/test_tools/test_integration.py: bafybeibfeepxlytrkrxhtruny4tgcfc2gw3hs5l3buydb2zxy3hvyj4ppq
+  tests/test_tools/base.py: bafybeia3sbvrvidzwgmqbqzqi5lbmhs7pnzcmcxbriduc7np5o552ckcle
+  tests/test_tools/test_base.py: bafybeigazjhuluth3ge2i6llh6zjtfvpdpd5x34vu3ctz7sfvmekc7vyce
+  tests/test_tools/test_common.py: bafybeifpbxupwbwfduivdd2igmxzy2m4gwkv3vmvgw65kjw6ngacihq2pm
+  tests/test_tools/test_integration.py: bafybeiesmfr2r5glo566qy5z3dtuqf6uckqskoamehhjpnjrsekbyjedjq
   tests/test_tools/test_rounds.py: bafybeicyz47ec4njsue67nltjv4fpr53d34qcjdr2rxipl5jbgbn3joybu
   tests/test_utils.py: bafybeigkob3r2tfsgazviaiwylmtyzhufodgrswf4byptqqm7k6uxdg5bq
   utils.py: bafybeiguanitvfc2ijitnwet33hvqxkztlhpe6nxfxamkj2z2bny5d3sjq
