+++ packages/valory/skills/abstract_round_abci/tests/test_tools/base.py
