+++ packages/valory/skills/abstract_round_abci/tests/test_tools/test_integration.py
@@ -43,7 +43,7 @@ from packages.valory.skills.abstract_round_abci.tests.data.dummy_abci.behaviours
 from packages.valory.skills.abstract_round_abci.tests.data.dummy_abci.rounds import (
     SynchronizedData,
 )
-from packages.valory.skills.abstract_round_abci.tests.test_tools.helpers import (
+from packages.valory.skills.abstract_round_abci.tests.test_tools.base import (
     FSMBehaviourTestToolSetup,
 )
